Restructure internal std folders
