+++ include/kernel/std/type_traits
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <type_traits_internal/concepts.hpp>
-#include <type_traits_internal/enums.hpp>
-#include <type_traits_internal/integral_constant.hpp>
-#include <type_traits_internal/type_manipulation.hpp>
+#include <internal/type_traits/concepts.hpp>
+#include <internal/type_traits/enums.hpp>
+#include <internal/type_traits/integral_constant.hpp>
+#include <internal/type_traits/type_manipulation.hpp>
